backend/pyproject.toml
[project]
name = "operatornext"
version = "0.1.0"
description = "LLM-Powered Browser Automation Agent"
authors = [
    { name = "HaiPro" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # 核心依赖
    "browser-use>=0.1.35",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.1",
    # AI & LLM
    "langchain>=0.3.14",
    "langchain-openai>=0.3.1",
    "langchain-anthropic>=0.3.3",
    "langchain-ollama>=0.2.2",
    "langchain-fireworks>=0.2.6",
    "langchain-aws>=0.2.11",
    "langchain-google-genai>=2.0.8",
    # 浏览器自动化
    "playwright>=1.49.0",
    # 数据处理
    "pydantic>=2.10.4",
    "beautifulsoup4>=4.12.3",
    # 工具
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "websockets>=12.0",
    "python-multipart>=0.0.7",
    "psutil>=6.1.1",
    # 数据库
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    # 安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "mangum>=0.17.0",
    "autogen-core>=0.4.9",
    "autogen-agentchat>=0.4.9",
    "autogen-ext[azure,openai]>=0.4.9",
    "pytest>=8.3.4",
    "requests>=2.32.3",
    "aiohttp>=3.11.11",
    # VM控制相关
    "docker>=7.0.0",
    "vncdotool>=1.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    # 开发工具
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "pre-commit>=3.6.0",
    
    # 测试
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    
    # 文档
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.3",
    
    # 调试和性能
    "tokencost>=0.1.16",
    "debugpy>=1.8.0"
]

[tool.ruff]
line-length = 88
lint.select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
lint.ignore = [
    "E501",  # 行太长
]
fix = true
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
